fix: change quota_remaining type from i64 to i32 to match DB INT4 column
The PostgreSQL column quota_remaining is INT4 (32-bit integer), but Rust
structs declared it as Option<i64> which maps to INT8. This caused sqlx
decode errors at runtime. Changed all quota-related types from i64 to i32
and SeaValue::BigInt to SeaValue::Int to match the actual DB schema.

+++ src/auth.rs
@@ -14,7 +14,7 @@ pub struct AuthResult {
     pub model_limits_enabled: bool,
     pub model_limits: Vec<String>,
     pub ip_whitelist: Vec<String>,
-    pub quota_remaining: Option<i64>,
+    pub quota_remaining: Option<i32>,
     pub quota_unlimited: bool,
 }
